Purpose of the United Kingdom
I think *the* main problem with the UK right now is that nobody knows what it is for.

Back round about 1950, a post-war goal of economic wealth through technology, along with longterm peace was made. In many ways we have now achieved that (despite its flaws). What is our goal now?

This is my proposal.