Provide "maternity box" for newborn babies
Modelled after the Finnish system. 
* http://www.kela.fi/web/en/maternitypackage
* http://www.buzzfeed.com/erinlarosa/heres-the-amazing-maternity-gift-box-all-new-parents-in-finl
* http://www.bbc.com/news/magazine-22751415

A key component of this — that isn't always recognised in articles about it — is that it's conditional on registering for pre-natal care. This is widely credited as a key reason for Finland having one of the lowest rates of infant mortality in the world (and almost half that of the UK at present).