Minutes etc should be published much quicker
Rather than allowing 24 hours for publishing of minutes/voting records etc., mandate that they should be published within 30 minutes. Changing to effectively real-time release is not only better for interested observers, but places the responsibility for doing it all with the meeting itself, rather than a separate function that happens later and can be thought about separately. This has precedent in Estonia.