Local schools
Replacing Secular Schools policy with Local Schools.
The problem with actively promoting the removal of Faith Schools is that there is a broad section of society that see value in Faith Schools and religious direction.  If this is to be a truly "people's manifesto" we need to work on mechanisms to promote more local influence in how educational establishments are set up and run; within a basic framework specified by the state.  Otherwise we risk the development of two distinct systems in the country - something that is happening to a certain extent right now with a Maintained and Academy/Free school programme running side by side.
Allowing for more local influence will naturally direct educational establishments towards the general consensus of the people - if, for example, the vast majority of local people do not have a religious belief then educational establishments should move in that direction (if governors/trustees are properly representative).
Closing down the last vestiges of Local Authorities and focusing on enhancing the Academies/Free School programme (and converting remaining Maintained schools) so that it does provide what is required in local communities reduces complexity and enables the state to focus on developing one system.