Provision for calling constitutional convention
A half-baked idea, but one I didn't want to have kicking round in my head. I'm not sure whether a constitutional convention should be how the constitution is first made, instead of a citizens' panel (not the same thing), or whether it should only exist for the purpose of making amendments.