Replacing the CoC and it's Concession of Rights
How to explain this.txt

Last time I was here I left in a huff because I held the view that contributing here is akin to signing all of your rights away. You can view that debate in issue #514 if you want. Essentially, my view on that hasn't changed and to me it appears to be a similar sort of situation to a Great Britain pre-magna carta. I don't like it very much.

I'd been thinking what to do about it:
I could fork the project. I think it would be useful somewhat, exactly how much I'm not sure. Nevertheless it's most certainly hard to do. It gives you some appreciation for the scale of the work that's already been put in.

I could have submitted an immediate repeal bill to request 514. Given that this was already voted for favourably and already sort-of on the books, it seemed to me to be counter-productive.

I could submit the alternative GNOME code of conduct. It gives you vague notions of what you are or aren't free to do, which is better than explicitly yielding those rights just to input anything into the project. But as an improvement, it's not much of one. It contains the sort of rules befitting a small project that exists just to keep things in check. If you have grandiose ambitions for the project or foresee it having millions of contributors, then you're going to need to get serious about it.

Which is what I've tried to do. I should point out that there should be a clause saying that this is not a legally-binding document, because it's got a lot of legalese in it. I try to avoid it (the simpler the words, the better) but it's hard to define these concepts outside of it.

Anyway... the last option is to submit an entirely new Code of Conduct. Great idea! Except that I'm struggling to write it. Constitutions are better as group projects. I like the scrutiny. So... this is why I'm submitting it, and this is why I'm submitting it unfinished. The reason I'm submitting it as a pull request, regardless of any vote outcome is because I don't want it to languish in the Issues/Ideas page. I want people to see it and to build on it.

I hope that explanation makes sense and I hope it eventually comes to something.